Evaluate a player's hand at showdown
Added a basic Factory and Collection as design patterns.
Replaced StandardSuitFactory with a new class StandardSuit.
Added support to instanciate Cards, Suits and CardCollections from strings.
Implemented HandEvaluator to identify the strength of each player's hand.
Added a new CLI Command to deal a hand to the Player and the board's community cards.