Determine and compare multiple hand strengths
Almost finished the HandStrengthCommand. It identifies which ranking
the player has, but it can not yet say what cards are part of that
ranking.

Added a new class HandStrength that contains information about the
strength of a certain poker hand.
Added a new class PlayerHand that extends CardCollection and has a
HandStrength.